Add clear, delete, and inputs to display
All the buttons are sent to the display. This is so the user can see
what is going on in the backend. They are also being stored in the code
through a local variable so it can be used later. A clear button was
added to reset the calculator. A delete button was added to remove the
last input in the input display.