Process inputs, helper functions, handle decimals
The buttons will automatically handle inputs so that way we don't get
anything messed up like multiple functions back to back. The inputString
will be formatted into an array with numbers and functions so that
way when it runs it will automatically when run through the operate
function. These are all hooked to run and get displayed when the equal
button is hit. Decimals are now only able to appear once. Several helper
functions were added to the code: checking if a value is a operation,
is numeric, converting a string to a number, and converting a string to
an operation function.