Add special equal mode
A special equals mode will be triggered once the equals button is hit.
If a user hits an operation, the input will become the result. If the
user hits a number, this will start a new string to solve. If the
user hits equal again, it will redo the last operation number pair.