Style, add glow light, and fix last input bug
Last input was grabbed at the beginning of the function but it should
only grab it the moment it is needed so the latest version is used. A
helper function for this purpose.  A glow light was added that triggered
when the decimal boolean is active or the equalsMode is active. There
has also been various CSS styling done to make the calculator feel more
Sci-Fi feeling.