Remove code, input object, and keyboard support
Some old code from the index.html was removed and renamed to be more
clear and match how the calculator actually functions as opposed to how
it was imagined at the beginning. An object called inputHandler will
manage all the calculations now. When an input is triggered, it can call
this object to run the necessary functions. Keyboard support was also
added in this commit using the newly added object. A small bug was also
fixed where the inputString would keep an operation without a number
pair in the inputString causing issues later.